+++ internal/mocks/maroto.go
@@ -61,12 +61,12 @@ func (_m *Maroto) Base64Image(base64 string, extension consts.Extension, prop ..
 }
 
 // Col provides a mock function with given fields: closure
-func (_m *Maroto) Col(width uint, closure func()) {
+func (_m *Maroto) Col(width float64, closure func()) {
 	_m.Called(closure)
 }
 
 // ColSpace provides a mock function with given fields:
-func (_m *Maroto) ColSpace(uint) {
+func (_m *Maroto) ColSpace(float64) {
 	_m.Called()
 }
